Split out functionality between absolute path and vanity path providers
